bug-fix/internal-issue-#549 (#9086)
* do not persist legacy analyzers into _analyzers table

* fix arangosearch upgrade in cluster

* get rid of Vasiliy's shit

* address review comments

* ensure link is synchronized after creation in upgrade

* fix compilation error

* minor cleanup

* fix tests

* distribute '_analyzers' collection as '_graphs'

* comment out Vasiliy's shit part 2

+++ tests/IResearch/IResearchOrder-test.cpp
@@ -182,7 +182,7 @@ void assertOrderExecutionFail(std::string const& queryString,
   return assertOrder(true, false, queryString, expected, exprCtx, bindVars, refName);
 }
 
-void assertOrderParseFail(std::string const& queryString, size_t parseCode) {
+void assertOrderParseFail(std::string const& queryString, int parseCode) {
   TRI_vocbase_t vocbase(TRI_vocbase_type_e::TRI_VOCBASE_TYPE_NORMAL, 1,
                         "testVocbase");
